move projectgroup
